diary: Moved application & configmanager module into the diary package.
