+++ diary/configmanager.py
