+++ main.py
@@ -1,9 +1,8 @@
 #!/usr/bin/env python3
 # coding: utf-8
 
-from application import App
+from diary.application import App
 
 
 app = App(name="Diary")
-app.set_conf("Test", "Testing Test")
